Created plot package and classes. Moved plot function from plot utils and changed usages.
