+++ tvb_epilepsy/scripts/seeg_data_scripts.py
@@ -1,8 +1,7 @@
 import numpy as np
 from scipy.signal import decimate
-
+from tvb_epilepsy.plot.plotter import Plotter
 from tvb_epilepsy.base.computations.analyzers_utils import filter_data
-from tvb_epilepsy.base.utils.plot_utils import plot_timeseries
 
 
 def prepare_seeg_observable(seeg_path, on_off_set, channels, win_len=5.0, low_freq=10.0, high_freq=None, log_flag=True,
@@ -10,15 +9,15 @@ def prepare_seeg_observable(seeg_path, on_off_set, channels, win_len=5.0, low_fr
     import re
     from pylab import detrend_linear
     from mne.io import read_raw_edf
-    from tvb_epilepsy.base.utils.plot_utils import plot_raster, plot_spectral_analysis_raster
     raw_data = read_raw_edf(seeg_path, preload=True)
     rois = np.where([np.in1d(s.split("POL ")[-1], channels) for s in raw_data.ch_names])[0]
     raw_data.resample(128.0)
     fs = raw_data.info['sfreq']
     data, times = raw_data[:, :]
     data = data[rois].T
+    plotter = Plotter()
     if plot_flag:
-        plot_spectral_analysis_raster(times, data, time_units="sec", freq=np.array(range(1,51,1)),
+        plotter.plot_spectral_analysis_raster(times, data, time_units="sec", freq=np.array(range(1,51,1)),
                                       title='Spectral Analysis', figure_name='Spectral Analysis', labels=channels,
                                       log_scale=True, save_flag=True)
     data_bipolar = []
@@ -36,11 +35,11 @@ def prepare_seeg_observable(seeg_path, on_off_set, channels, win_len=5.0, low_fr
     data_filtered = np.array(data_filtered).T
     # filter_data, times = raw_data.filter(low_freq, 100.0, picks=rois)[:, :]
     if plot_flag:
-        plot_spectral_analysis_raster(times, data_bipolar, time_units="sec", freq=np.array(range(1, 51, 1)),
+        plotter.plot_spectral_analysis_raster(times, data_bipolar, time_units="sec", freq=np.array(range(1, 51, 1)),
                                   title='Spectral Analysis',
                                   figure_name='Spectral Analysis Bipolar', labels=bipolar_channels,
                                   show_flag=True, save_flag=True, log_scale=True)
-        plot_spectral_analysis_raster(times, data_filtered, time_units="sec", freq=np.array(range(1, 51, 1)),
+        plotter.plot_spectral_analysis_raster(times, data_filtered, time_units="sec", freq=np.array(range(1, 51, 1)),
                                   title='Spectral Analysis_bipolar', figure_name='Spectral Analysis Filtered', labels=bipolar_channels,
                                       log_scale=True, save_flag=True)
     del data
@@ -66,7 +65,7 @@ def prepare_seeg_observable(seeg_path, on_off_set, channels, win_len=5.0, low_fr
     times = times[t_onset:t_offset]
     observation = observation[t_onset:t_offset]
     if plot_flag:
-        plot_raster(times, {"observation": observation}, time_units="sec", special_idx=None, title='Time Series',
+        plotter.plot_raster(times, {"observation": observation}, time_units="sec", special_idx=None, title='Time Series',
                     offset=1.0, figure_name='TimeSeries', labels=bipolar_channels, save_flag=True)
     # n_times = times.shape[0]
     # observation = resample_poly(observation, 2048, n_times)
@@ -75,6 +74,6 @@ def prepare_seeg_observable(seeg_path, on_off_set, channels, win_len=5.0, low_fr
     observation -= observation.min()
     observation /= observation.max()
     if plot_flag:
-        plot_timeseries(times, {"observation": observation}, time_units="sec", special_idx=None, title='Time Series',
+        plotter.plot_timeseries(times, {"observation": observation}, time_units="sec", special_idx=None, title='Time Series',
                     figure_name='TimeSeriesDecimated', labels=bipolar_channels, show_flag=True, save_flag=True) #
     return observation, times, fs/2
